requirements.txt
notebook
aiohttp
ipywidgets
pytest
nltk
tensorflow
pyspark
jsonlines
langdetect
indic-nlp-library
